[cpackget] add command shall download the pack and extract the license files to .Download #196
cpackget is already prepared to just extract the EULA and then quit, and to auto-accept EULA in a next step.
Changed the program flow to skip showing the license / preparing the license UI if already the command line specifies the extract option.

If the license file already exists it will be deleted (error if not possible).

Added for encoded progress:
* L: License file follows
I: [L:F"C:\Users\user\AppData\Local\Arm\Packs\.Download\Keil.MDK-Middleware_Graphics.1.3.0.pack.license.rtf"]

+++ cmd/installer/pack.go
@@ -14,6 +14,7 @@ import (
 	"strings"
 
 	"github.com/lu4p/cat"
+	"github.com/open-cmsis-pack/cpackget/cmd/errors"
 	errs "github.com/open-cmsis-pack/cpackget/cmd/errors"
 	"github.com/open-cmsis-pack/cpackget/cmd/ui"
 	"github.com/open-cmsis-pack/cpackget/cmd/utils"
@@ -460,7 +461,20 @@ func (p *PackType) extractEula(packPath string) error {
 
 	eulaFileName := packPath + "." + path.Base(p.Pdsc.License)
 
-	log.Infof("Extracting embedded license to %v", eulaFileName)
+	if utils.GetEncodedProgress() {
+		log.Infof("[L:F\"%s\"]", eulaFileName)
+	} else {
+		log.Infof("Extracting embedded license to %v", eulaFileName)
+	}
+
+	if utils.FileExists(eulaFileName) {
+		utils.UnsetReadOnly(eulaFileName)
+		os.Remove(eulaFileName)
+	}
+	if utils.FileExists(eulaFileName) {
+		log.Errorf("Cannot remove previous copy of license file: \"%s\"", eulaFileName)
+		return errors.ErrFailedCreatingFile
+	}
 
 	return os.WriteFile(eulaFileName, eulaContents, utils.FileModeRO)
 }
